Changed data type again

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,15 +9,15 @@
 
 int main(void)
 {
-	double first = 0;
-	double second = 1;
-	double result = 0;
+	long double first = 0;
+	long double second = 1;
+	long double result = 0;
 	int i;
 
 	for (i = 0; i < 98; i++)
 	{
 		result = first + second;
-		printf("%.0lf", result);
+		printf("%.0Lf", result);
 		if (i != 97)
 			printf(", ");
 		first = second;
